Copy sizeDistribution() over from maxEnt_method.py; started the sizeDistribution() conversion into a class, sizeDistribution; Removed poresize.py and renamed to SizeDistribution.py.
